requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
nemo-toolkit==1.20.0
torch==2.1.0
torchaudio==2.1.0
onnx==1.14.1
onnxruntime-gpu==1.16.1
python-multipart==0.0.6
pydantic==2.4.2
numpy>=1.22,<1.24
soundfile>=0.12.1